pyproject.toml
[project]
name = "ainm2026"
version = "0.1.0"
description = "Norwegian AI Championship 2026 - Tripletex Accounting Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # --- Serving ---
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "urllib3",
    "httpx>=0.28.1",
    # --- PDF parsing ---
    "numpy>=2.3.5",
    "pymupdf>=1.27.2",
    # --- LLM ---
    "langchain>=1.2.10",
    "langchain-core>=1.2.16",
    "langchain-openai>=1.1.10",
    "langchain-google-genai>=2.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tasks", "api"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
markers = [
    "sandbox: requires TRIPLETEX_BASE_URL and TRIPLETEX_SESSION_TOKEN env vars",
]
